I wanted to see if there was an easy way to use all training examples and the reward function to get a stable learning alogirthm. I was not able to succeed.
The main issue is that because my reward is always positive and I am encouraging gradient ascent steps to just increase towards the current operations. It is just optimizing to have more and more accurate predictions to match earlier outcomes.

I need to think of a better way to combine the different functions