this did not work.
I was trying to only train the network on episodes that went longer than a certain time period. The hope was that it would learn what these networks were doing well and learn those values. What I found instead was that it basically learned to only go one direction at the end presumably because I was telling it the example cases were correct with similar cross entropy cost function and all of the examples failed in the end. It did not realize that it wanted to do better than systematically fail towards the end.

I was getting pi values for the two actions in cart pole like the following:
Entropy input:  [[2.3831616e-01 7.6168382e-01]
 [4.7573900e-01 5.2426094e-01]
 [7.2258633e-01 2.7741367e-01]
 [4.8678452e-01 5.1321542e-01]
 [7.3212337e-01 2.6787660e-01]
 [5.0059533e-01 4.9940470e-01]
 [2.7041581e-01 7.2958410e-01]
 [1.2395370e-01 8.7604636e-01]
 [2.8909466e-01 7.1090537e-01]
 [5.4213411e-01 4.5786586e-01]
 [3.0234945e-01 6.9765055e-01]
 [5.5807686e-01 4.4192320e-01]
 [3.1630281e-01 6.8369716e-01]
 [1.4830467e-01 8.5169536e-01]
 [6.4370260e-02 9.3562973e-01]
 [2.8186485e-02 9.7181356e-01]
 [6.9658995e-02 9.3034106e-01]
 [1.6942954e-01 8.3057046e-01]
 [7.2038822e-02 9.2796123e-01]
 [3.0667976e-02 9.6933204e-01]
 [1.3786635e-02 9.8621339e-01]
 [6.6899257e-03 9.9331009e-01]
 [3.5269307e-03 9.9647301e-01]
 [2.0181821e-03 9.9798179e-01]
 [1.2477082e-03 9.9875224e-01]
 [8.2824798e-04 9.9917173e-01]
 [5.8639218e-04 9.9941361e-01]]