figured out what was wrong with my version by doing a large amount of copying of a version from the internet stored in cartpole_reinforce.py
The main issue was that I was taking the new state and setting it directly into state before
i had added the state to my state list. So when I was doing gradient descent I was comparing the action taken in t with the state used in t + 1. This mismatch was leading to gradient descent that did not go to the optimum value.